# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ads-mrkt_internal_market_application_market_http_model.CreateListingRequest:
    properties:
      categories:
        items:
          type: string
        type: array
      channel_id:
        type: integer
      description:
        type: string
      prepared_post:
        items:
          type: integer
        type: array
      prices:
        items:
          type: integer
        type: array
      status:
        type: string
      type:
        type: string
    type: object
...
  ads-mrkt_internal_market_application_market_http_model.UpdateListingRequest:
    properties:
      categories:
        items:
          type: string
        type: array
      description:
        type: string
      prepared_post:
        items:
          type: integer
        type: array
      prices:
        items:
          type: integer
        type: array
      status:
        type: string
      type:
        type: string
    type: object
...
  ads-mrkt_internal_market_domain_entity.Listing:
    properties:
      categories:
        description: JSON array of strings from predefined set
        items:
          type: integer
        type: array
      channel_followers:
        type: integer
      channel_id:
        type: integer
      channel_photo:
        type: string
      channel_title:
        type: string
      channel_username:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      prepared_post:
        description: 'JSON object, e.g. {"message": "ad text"}; extensible'
        items:
          type: integer
        type: array
      prices:
        items:
          type: integer
        type: array
      status:
        $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.ListingStatus'
      type:
        $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.ListingType'
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  ads-mrkt_internal_market_domain_entity.ListingStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - ListingStatusActive
    - ListingStatusInactive
  ads-mrkt_internal_market_domain_entity.ListingType:
    enum:
    - lessor
    - lessee
    type: string
    x-enum-varnames:
    - ListingTypeLessor
    - ListingTypeLessee
...
  ads-mrkt_internal_server_templates_response.Template:
    properties:
      data: {}
      error_code:
        type: string
      ok:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger for Ads Market API
  version: "1.0"
paths:
...
  /market/listings:
    get:
      parameters:
      - description: 'Filter by type: lessor | lessee'
        in: query
        name: type
        type: string
      - description: Comma-separated categories (e.g. Tech,Crypto)
        in: query
        name: categories
        type: string
      - description: Min channel followers (only lessor listings with stats)
        in: query
        name: min_followers
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of listings
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.Listing'
                  type: array
              type: object
      summary: List all listings with optional type, categories, and min_followers
        filter (public, no auth)
      tags:
      - Market
    post:
      consumes:
      - application/json
      parameters:
      - description: listing body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.CreateListingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created listing
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.Listing'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Create listing
      tags:
      - Market
  /market/listings/{id}:
    delete:
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Delete listing
      tags:
      - Market
    get:
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Listing
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.Listing'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      summary: Get listing by ID
      tags:
      - Market
    patch:
      consumes:
      - application/json
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: integer
      - description: fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ads-mrkt_internal_market_application_market_http_model.UpdateListingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated listing
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.Listing'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not found
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Update listing
      tags:
      - Market
...
  /market/my-listings:
    get:
      parameters:
      - description: 'Filter by type: lessor | lessee'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of my listings
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ads-mrkt_internal_market_domain_entity.Listing'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/ads-mrkt_internal_server_templates_response.Template'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: List current user's listings with optional type filter
      tags:
      - Market
...
schemes:
- http
security:
- JWT: []
securityDefinitions:
  JWT:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"